+++ src/components/button-show.js
@@ -1,0 +1,5 @@
+export const creatButtonShowMoreTemplate = () => {
+  return (
+    `<button class="films-list__show-more">Show more</button>`
+  );
+};
