+++ src/components/film-card-list.js
@@ -1,0 +1,9 @@
+import {creatFilmCardTemplate} from "./film-card.js";
+
+export const getListCards = (count) => {
+  let listCards = ``;
+  for (let i = 0; i < count; i++) {
+    listCards += creatFilmCardTemplate();
+  }
+  return listCards;
+};
