+++ src/components/footer-statistics.js
@@ -1,0 +1,5 @@
+export const creatFooterStatisticsTemplate = () => {
+  return (
+    `<p>130 291 movies inside</p>`
+  );
+};
